Added tooltips for nodes, runs dropdown is ready.
