+++ frontend/src/components/CustomNodesContent/CheckpointNode/index.jsx
@@ -1,6 +1,6 @@
 import { faMapMarkedAlt } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { Grid, Typography } from '@mui/material';
+import { Grid, Tooltip, Typography } from '@mui/material';
 import { Box } from '@mui/system';
 import { useEffect, useState } from 'react';
 import { Handle, Position } from 'react-flow-renderer';
@@ -44,28 +44,30 @@ const CheckpointNode = (props) => {
     }, [RunState[props.id].get()]);
 
     return (
-        <Box sx={{ ...customNodeStyle, border: `3px solid ${borderColor}` }}>
-            <Handle type="source" position={Position.Right} id="checkpoint_source" style={{ backgroundColor: 'red', height: 10, width: 10 }} />
-            <Handle type="target" position={Position.Left} id="checkpoint_target" style={{ backgroundColor: 'red', height: 10, width: 10 }} />
-            <Grid container alignItems="flex-start" wrap="nowrap">
-                <Box component={FontAwesomeIcon} fontSize={19} color="secondary.main" icon={faMapMarkedAlt} />
-                <Grid item ml={1.5} textAlign="left">
-                    <Typography fontSize={11} fontWeight={900}>
-                        Checkpoint
-                    </Typography>
-                </Grid>
-
-                {isEditorPage && (
-                    <Grid position="absolute" bottom={2} right={9} container wrap="nowrap" width="auto" alignItems="center" justifyContent="space-between">
-                        <Box mt={2}>
-                            <MoreInfoMenu iconHorizontal iconColor="#0073C6" iconColorDark="#0073C6" iconSize={19} noPadding>
-                                <PlayTriggerNodeItem />
-                            </MoreInfoMenu>
-                        </Box>
+        <Tooltip title={'Node ID: ' + props.id} placement="top">
+            <Box sx={{ ...customNodeStyle, border: `3px solid ${borderColor}` }}>
+                <Handle type="source" position={Position.Right} id="checkpoint_source" style={{ backgroundColor: 'red', height: 10, width: 10 }} />
+                <Handle type="target" position={Position.Left} id="checkpoint_target" style={{ backgroundColor: 'red', height: 10, width: 10 }} />
+                <Grid container alignItems="flex-start" wrap="nowrap">
+                    <Box component={FontAwesomeIcon} fontSize={19} color="secondary.main" icon={faMapMarkedAlt} />
+                    <Grid item ml={1.5} textAlign="left">
+                        <Typography fontSize={11} fontWeight={900}>
+                            Checkpoint
+                        </Typography>
                     </Grid>
-                )}
-            </Grid>
-        </Box>
+
+                    {isEditorPage && (
+                        <Grid position="absolute" bottom={2} right={9} container wrap="nowrap" width="auto" alignItems="center" justifyContent="space-between">
+                            <Box mt={2}>
+                                <MoreInfoMenu iconHorizontal iconColor="#0073C6" iconColorDark="#0073C6" iconSize={19} noPadding>
+                                    <PlayTriggerNodeItem />
+                                </MoreInfoMenu>
+                            </Box>
+                        </Grid>
+                    )}
+                </Grid>
+            </Box>
+        </Tooltip>
     );
 };
