+++ frontend/src/components/CustomNodesContent/ScheduleNode/index.jsx
@@ -1,6 +1,6 @@
 import { faClock } from '@fortawesome/free-regular-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { Grid, Typography } from '@mui/material';
+import { Grid, Tooltip, Typography } from '@mui/material';
 import { Box } from '@mui/system';
 import { useEffect, useState } from 'react';
 import { Handle } from 'react-flow-renderer';
@@ -44,31 +44,33 @@ const ScheduleNode = (props) => {
     }, [RunState[props.id].get()]);
 
     return (
-        <Box sx={{ ...customNodeStyle, border: `3px solid ${borderColor}` }}>
-            <Handle type="source" position="right" id="schedule" style={{ backgroundColor: 'red', height: 10, width: 10 }} />
-            <Grid container alignItems="flex-start" wrap="nowrap">
-                <Box component={FontAwesomeIcon} fontSize={19} color="secondary.main" icon={faClock} />
-                <Grid item ml={1.5} textAlign="left">
-                    <Typography fontSize={11} fontWeight={900}>
-                        Schedule trigger
-                    </Typography>
+        <Tooltip title={'Node ID: ' + props.id} placement="top">
+            <Box sx={{ ...customNodeStyle, border: `3px solid ${borderColor}` }}>
+                <Handle type="source" position="right" id="schedule" style={{ backgroundColor: 'red', height: 10, width: 10 }} />
+                <Grid container alignItems="flex-start" wrap="nowrap">
+                    <Box component={FontAwesomeIcon} fontSize={19} color="secondary.main" icon={faClock} />
+                    <Grid item ml={1.5} textAlign="left">
+                        <Typography fontSize={11} fontWeight={900}>
+                            Schedule trigger
+                        </Typography>
 
-                    <Typography fontSize={10} mt={1}>
-                        Every 5 minutes
-                    </Typography>
+                        <Typography fontSize={10} mt={1}>
+                            Every 5 minutes
+                        </Typography>
+                    </Grid>
                 </Grid>
-            </Grid>
 
-            {isEditorPage && (
-                <Grid position="absolute" bottom={2} right={9} container wrap="nowrap" width="auto" alignItems="center" justifyContent="space-between">
-                    <Box mt={2}>
-                        <MoreInfoMenu iconHorizontal iconColor="#0073C6" iconColorDark="#0073C6" iconSize={19} noPadding>
-                            <ScheduleTriggerNodeItem />
-                        </MoreInfoMenu>
-                    </Box>
-                </Grid>
-            )}
-        </Box>
+                {isEditorPage && (
+                    <Grid position="absolute" bottom={2} right={9} container wrap="nowrap" width="auto" alignItems="center" justifyContent="space-between">
+                        <Box mt={2}>
+                            <MoreInfoMenu iconHorizontal iconColor="#0073C6" iconColorDark="#0073C6" iconSize={19} noPadding>
+                                <ScheduleTriggerNodeItem />
+                            </MoreInfoMenu>
+                        </Box>
+                    </Grid>
+                )}
+            </Box>
+        </Tooltip>
     );
 };
