+++ frontend/src/pages/View/index.jsx
@@ -1,34 +1,26 @@
 import { useTheme } from '@emotion/react';
-import { faPlayCircle } from '@fortawesome/free-regular-svg-icons';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { Box, Button, Drawer, Grid, IconButton, MenuItem, TextField, Typography } from '@mui/material';
+import { Box, Button, Drawer, Grid, Typography } from '@mui/material';
 import { useSnackbar } from 'notistack';
 import { useEffect, useRef, useState } from 'react';
 import ReactFlow, { addEdge, Controls } from 'react-flow-renderer';
 import { useHistory, useParams, useLocation } from 'react-router-dom';
-import { Downgraded } from '@hookstate/core';
 import CustomChip from '../../components/CustomChip';
 import CustomLine from '../../components/CustomNodesContent/CustomLine';
 import PublishPipelineDrawer from '../../components/DrawerContent/PublishPipelineDrawer';
 import RemoveLogsPageItem from '../../components/MoreInfoContent/RemoveLogsPageItem';
 import MoreInfoMenu from '../../components/MoreInfoMenu';
 import { useGetPipelineFlow } from '../../graphql/getPipelineFlow';
-import { useRunPipelines } from '../../graphql/runPipelines';
-import { useStopPipelines } from '../../graphql/stopPipelines';
-import { usePipelineTasksRun } from '../../graphql/getPipelineTasksRun';
 import { edgeTypes, nodeTypes, useGlobalFlowState } from '../Flow';
 import { useGlobalEnvironmentState } from '../../components/EnviromentDropdown';
-import useWebSocket, { useGlobalRunState } from './useWebSocket';
+import RunsDropdown from './RunsDropdown';
+import Timer from './Timer';
 
 const View = () => {
     // Hooks
     const theme = useTheme();
     const history = useHistory();
     const { state: pipeline } = useLocation();
-    const getPipelineFlow = useGetPipelineFlow_(pipeline);
-    const runPipelines = useRunPipelines_();
-    const stopPipelines = useStopPipelines_();
-    const getPipelineTasksRun = usePipelineTasksRunHook();
+    const getPipelineFlow = useGetPipelineFlowHook(pipeline);
 
     // URI parameter
     const { pipelineId } = useParams();
@@ -38,12 +30,10 @@ const View = () => {
 
     // Page states
     const [isOpenPublishDrawer, setIsOpenPublishDrawer] = useState(false);
-    const [isRunning, setIsRunning] = useState(false);
-    const [runID, setRunID] = useState('');
     const [, setIsLoadingFlow] = useState(true);
-    // const [timeElapsed, setTimeElapsed] = useState(Math.floor((new Date() - FlowState.startedRunningAt.attach(Downgraded).get()) / 1000));
-    const [timeElapsed, setTimeElapsed] = useState(0);
-    const [tasks, setTasks] = useState();
+    // // const [timeElapsed, setTimeElapsed] = useState(Math.floor((new Date() - FlowState.startedRunningAt.attach(Downgraded).get()) / 1000));
+    // const [timeElapsed, setTimeElapsed] = useState(0);
+    // const [tasks, setTasks] = useState();
 
     const Environment = useGlobalEnvironmentState();
 
@@ -70,15 +60,6 @@ const View = () => {
         // eslint-disable-next-line react-hooks/exhaustive-deps
     }, []);
 
-    // Get current runs status
-    useEffect(() => {
-        if (isRunning && runID !== '') {
-            getPipelineTasksRun(runID, Environment.id.get());
-        }
-
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [isRunning, runID]);
-
     // Handle edit button
     const handleGoToEditorPage = () => {
         FlowState.isEditorPage.set(true);
@@ -89,15 +70,6 @@ const View = () => {
     const onLoad = (_reactFlowInstance) => setReactFlowInstance(_reactFlowInstance);
     const onConnect = (params) => {
         setElements((els) => addEdge({ ...params, type: 'custom' }, els));
-    };
-
-    // Instantiate websocket connection
-    const socketResponse = useWebSocket(Environment.id.get(), runID);
-
-    const handleTimerStart = () => {
-        FlowState.isRunning.set(true);
-        setIsRunning(true);
-        runPipelines(Environment.id.get(), setRunID);
     };
 
     // // Updates timer every second
@@ -166,43 +138,11 @@ const View = () => {
                         <CustomChip amount={2} label="Workers online" margin={2} customColor="purple" />
                     </Grid>
 
-                    <Grid item alignItems="center" display="flex" flex={1}>
-                        <Typography variant="h3">Run</Typography>
-                        <TextField label="Live" id="last" select size="small" sx={{ ml: 2, mr: 2, flex: 1 }}>
-                            {FlowState.isRunning.get() ? (
-                                <MenuItem value="live">
-                                    {formatDate(FlowState.startedRunningAt.attach(Downgraded).get())} - {runID}
-                                </MenuItem>
-                            ) : null}
-                            <MenuItem value="24">Last 24 hours</MenuItem>
-                        </TextField>
-                    </Grid>
-                    <Grid item flex={0.6}>
-                        {isRunning ? (
-                            <Box display="flex" alignItems="center" ml={2}>
-                                <Button
-                                    onClick={() => {
-                                        FlowState.isRunning.set(false);
-                                        setIsRunning(false);
-                                        stopPipelines(Environment.id.get(), runID);
-                                    }}
-                                    variant="outlined"
-                                    color="error"
-                                    sx={{ width: 70, fontWeight: '700', fontSize: '.81rem', border: 2, '&:hover': { border: 2 } }}>
-                                    Stop
-                                </Button>
+                    {/* Runs dropdown */}
+                    <RunsDropdown />
 
-                                <Typography variant="h3" ml={2}>
-                                    00:00:00
-                                    {/* 00:00:{timeElapsed} */}
-                                </Typography>
-                            </Box>
-                        ) : (
-                            <IconButton sx={{ margin: 0, height: 40, width: 40 }} aria-label="play" id="play-button" aria-haspopup="true" onClick={handleTimerStart}>
-                                <Box component={FontAwesomeIcon} fontSize={30} sx={{ color: 'cyan.main' }} icon={faPlayCircle} />
-                            </IconButton>
-                        )}
-                    </Grid>
+                    {/* Timer */}
+                    <Timer environmentID={Environment.id.get()} />
                 </Grid>
             </Box>
 
@@ -244,7 +184,7 @@ const LOGS_MOCK = {
 export default View;
 
 // ------ Custom hooks
-const useGetPipelineFlow_ = (pipeline) => {
+const useGetPipelineFlowHook = (pipeline) => {
     // GraphQL hook
     const getPipelineFlow = useGetPipelineFlow();
 
@@ -275,84 +215,6 @@ const useGetPipelineFlow_ = (pipeline) => {
         } else {
             setElements(response);
             FlowState.elements.set(response);
-        }
-    };
-};
-
-const useRunPipelines_ = () => {
-    // GraphQL hook
-    const runPipelines = useRunPipelines();
-
-    // URI parameter
-    const { pipelineId } = useParams();
-
-    const { enqueueSnackbar, closeSnackbar } = useSnackbar();
-
-    // Run pipeline flow
-    return async (environmentID, setRunID) => {
-        const response = await runPipelines({ pipelineID: pipelineId, environmentID });
-
-        if (response.r === 'error') {
-            closeSnackbar();
-            enqueueSnackbar("Can't run flow: " + response.msg, { variant: 'error' });
-        } else if (response.errors) {
-            response.errors.map((err) => enqueueSnackbar(err.message + ': run flow', { variant: 'error' }));
-        } else {
-            setRunID(response.run_id);
-            enqueueSnackbar('Success', { variant: 'success' });
-        }
-    };
-};
-
-const useStopPipelines_ = () => {
-    // GraphQL hook
-    const stopPipelines = useStopPipelines();
-
-    // URI parameter
-    const { pipelineId } = useParams();
-
-    const { enqueueSnackbar, closeSnackbar } = useSnackbar();
-
-    // Stop pipeline flow
-    return async (environmentID, runID) => {
-        const response = await stopPipelines({ pipelineID: pipelineId, environmentID, runID });
-
-        if (response.r === 'error') {
-            closeSnackbar();
-            enqueueSnackbar("Can't stop flow: " + response.msg, { variant: 'error' });
-        } else if (response.errors) {
-            response.errors.map((err) => enqueueSnackbar(err.message + ': stop flow', { variant: 'error' }));
-        } else {
-            enqueueSnackbar('Success', { variant: 'success' });
-        }
-    };
-};
-
-const usePipelineTasksRunHook = () => {
-    // GraphQL hook
-    const getPipelineTasksRun = usePipelineTasksRun();
-
-    // URI parameter
-    const { pipelineId } = useParams();
-
-    const RunState = useGlobalRunState();
-
-    const { enqueueSnackbar, closeSnackbar } = useSnackbar();
-
-    // Update pipeline flow
-    return async (runID, environmentID) => {
-        // Prepare input to match the structure in the backend
-
-        const response = await getPipelineTasksRun({ pipelineID: pipelineId, runID, environmentID });
-
-        if (response.r === 'Unauthorized') {
-            closeSnackbar();
-            enqueueSnackbar(`Can't update flow: ${response.r}`, { variant: 'error' });
-        } else if (response.errors) {
-            response.errors.map((err) => enqueueSnackbar(err.message + ': update flow failed', { variant: 'error' }));
-        } else {
-            response.map((a) => RunState[a.node_id].set(a.status));
-            // enqueueSnackbar('Success', { variant: 'success' });
         }
     };
 };
@@ -394,11 +256,4 @@ function prepareInputForFrontend(input) {
     }
 
     return [...edgesInput, ...nodesInput];
-}
-
-export function formatDate(date) {
-    let day = new Intl.DateTimeFormat('en', { day: 'numeric' }).format(date);
-    let monthYear = new Intl.DateTimeFormat('en', { year: 'numeric', month: 'short' }).format(date);
-    let time = new Intl.DateTimeFormat('en', { hourCycle: 'h23', hour: '2-digit', minute: 'numeric', second: 'numeric' }).format(date);
-    return `${day} ${monthYear} ${time}`;
 }
